pyproject.toml
[project]
name = "oelp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "chromadb>=1.1.0",
    "fastapi>=0.116.1",
    "fitz>=0.0.1.dev2",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.2",
    "passlib>=1.7.4",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "razorpay>=2.0.0",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "stripe>=12.5.1",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
    "bcrypt>=4.3.0",
]
